+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Now using Java 7 source level
 * Bumped all plugins / dependencies to latest versions
+* Moved all `slick2d-core` tests in `src/test/` and added test data. Test classes are not shipped any more in the main JAR
 
 ## v1.0.0 (2014-04-16)